+++ src/main/java/employee_csv_project/controller/db_controller/db_management/ConnectionManager.java
@@ -1,4 +1,4 @@
-package employee_csv_project.controller.db_management;
+package employee_csv_project.controller.db_controller.db_management;
 
 import employee_csv_project.config.Config;
